+++ RoutineGlare.h
@@ -15,7 +15,7 @@ class CRoutineGlare : public CRoutine
                       CRGB         base_color,
                       size_t       q,
                       bool         forward,
-                      uint32_t     period_sec);
+                      size_t     period_sec);
         ~CRoutineGlare();
 
     public:
@@ -26,14 +26,14 @@ class CRoutineGlare : public CRoutine
         CRGB     BaseColorRGB() { return CRGB(m_base_color); }
         CHSV     BaseColor()    { return m_base_color; }
         bool     Forward()      { return m_forward; }
-        uint32_t PeriodSec()    { return m_period_sec; }
+        size_t PeriodSec()    { return m_period_sec; }
 
     private:
         CHSV         m_base_color;
         size_t       m_q;
         bool         m_forward    = true;
-        uint32_t     m_period_sec = 10;
-        uint32_t     m_last_run   = 0;
+        size_t     m_period_sec = 10;
+        size_t     m_last_run   = 0;
         float        m_midpoint   = 0.0;
 
     private:
