+++ RoutineSticks.h
@@ -10,8 +10,8 @@ class CRoutineSticks : public CRoutine
     public:
         static constexpr size_t   c_alloc_qty      = 1;
         static constexpr size_t   c_max_num_sticks = 32;
-        static constexpr uint32_t c_min_period_sec = 20;
-        static constexpr uint32_t c_max_period_sec = 40;
+        static constexpr size_t c_min_period_sec = 20;
+        static constexpr size_t c_max_period_sec = 40;
         static constexpr float   c_min_width_pct  = 0.02;
         static constexpr float   c_max_width_pct  = 0.10;
